Always need to cast argument to isspace to int
git-svn-id: https://svn.code.sf.net/p/moonunit/svn/trunk/moonunit@178 7cfde7b9-103a-0410-917d-a4b0b378e324

+++ src/moonunit/upopt.c
@@ -364,7 +364,7 @@ Upopt_PrintHelp(UpoptContext* context, FILE* out, int columns)
                 
                 if (ws)
                 {
-                    while (isspace(*ws)) ws++;
+                    while (isspace((int) *ws)) ws++;
                     cur = *ws ? ws : NULL;
                 }
                 else
